+++ src/main/groovy/io/xh/hoist/configuration/RuntimeConfig.groovy
@@ -17,10 +17,40 @@ import static io.xh.hoist.util.Utils.withDelegate
  */
 class RuntimeConfig {
 
+    /**
+     * All apps should call this from runtime.groovy
+     * to setup necessary default configurations
+     */
     static void defaultConfig(Script script) {
         withDelegate(script) {
             grails.serverURL = getInstanceConfig('serverURL')
         }
     }
 
+    /**
+     * Call this from runtime.groovy
+     * to setup an in memory H2 DB instead of
+     * a MySQL or SQL Server DB.  This H2 DB option
+     * is intended only for early stages of development,
+     * before a production ready DB has been set up.
+     */
+    static void h2Config(Script script) {
+        withDelegate(script) {
+            dataSource {
+                pooled = true
+                jmxExport = true
+                driverClassName = "org.h2.Driver"
+                username = "sa"
+                password = ""
+            }
+            environments {
+                development {
+                    dataSource {
+                        dbCreate = "create-drop"
+                        url = "jdbc:h2:mem:devDb;LOCK_TIMEOUT=10000;DB_CLOSE_ON_EXIT=FALSE"
+                    }
+                }
+            }
+        }
+    }
 }